requirements.txt
boto3
tweepy
kafka-python
beautifulsoup4
requests
transformers
torch
flask
pandas
nltk
